Rename internal math.{h,cpp} to mathml.{h,cpp}.
