+++ src/mathml.cpp
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#include "math.h"
+#include "mathml.h"
 
 namespace libcellml {
